issue #605: refactoring (DRY)
