+++ proarc-common/src/main/java/cz/cas/lib/proarc/common/mods/ndk/RdaNdkMapper.java
@@ -1,0 +1,65 @@
+/*
+ * Copyright (C) 2017 Martin Rumanek
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package cz.cas.lib.proarc.common.mods.ndk;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import cz.cas.lib.proarc.common.mods.custom.ModsConstants;
+import cz.cas.lib.proarc.common.object.ndk.NdkMetadataHandler;
+import cz.cas.lib.proarc.mods.ModsDefinition;
+import cz.cas.lib.proarc.mods.RecordInfoDefinition;
+import cz.cas.lib.proarc.mods.StringPlusLanguagePlusAuthority;
+import java.io.IOException;
+
+public class RdaNdkMapper extends NdkMapper {
+
+    @Override
+    public final RdaModsWrapper toJsonObject(ModsDefinition mods, NdkMapper.Context ctx) {
+        NdkMapper.RdaModsWrapper wrapper = new NdkMapper.RdaModsWrapper();
+        wrapper.setMods(mods);
+        if (mods.getRecordInfo().isEmpty() || mods.getRecordInfo().get(0).getDescriptionStandard().isEmpty()) {
+            return wrapper;
+        }
+        String descriptionStandard = mods.getRecordInfo().get(0).getDescriptionStandard().get(0).getValue();
+        mods.getRecordInfo().get(0).getDescriptionStandard().clear();
+        if (descriptionStandard.equalsIgnoreCase(ModsConstants.VALUE_DESCRIPTIONSTANDARD_RDA)) {
+            wrapper.setRdaRules(true);
+        } else {
+            wrapper.setRdaRules(false);
+        }
+        return wrapper;
+    }
+
+    @Override
+    public ModsDefinition fromJsonObject(ObjectMapper jsMapper, String json, Context ctx) throws IOException {
+        RdaModsWrapper wrapper = jsMapper.readValue(json, RdaModsWrapper.class);
+        ModsDefinition mods = wrapper.getMods();
+        StringPlusLanguagePlusAuthority descriptionStandard = new StringPlusLanguagePlusAuthority();
+        if (wrapper.getRdaRules() != null && wrapper.getRdaRules()) {
+            descriptionStandard.setValue(ModsConstants.VALUE_DESCRIPTIONSTANDARD_RDA);
+        } else {
+            descriptionStandard.setValue(ModsConstants.VALUE_DESCRIPTIONSTANDARD_AACR);
+        }
+        if (mods.getRecordInfo().isEmpty()) {
+            RecordInfoDefinition recordInfo = new RecordInfoDefinition();
+            recordInfo.getDescriptionStandard().add(0, descriptionStandard);
+            mods.getRecordInfo().add(0, recordInfo);
+        } else {
+            mods.getRecordInfo().get(0).getDescriptionStandard().add(0, descriptionStandard);
+        }
+        return mods;
+    }
+}
